=Improved exception handling, reformed login/logout, added helpful annotatons

+++ mips/views.py
@@ -17,61 +17,87 @@ def Authenticated(view):
         if user:
             return view(user, *args, **kwargs)
         else:
-            return login(None)
+            return index(None)
     return f
 
 def index(request):
     logged_in = users.get_current_user() != None
-    return render_to_response("mips/index.html", {'logged_in' : logged_in})
-
-def login(request):
-    return render_to_response("mips/login.html",
-                              {"login_url": users.create_login_url("/programs/")})
-
-def logout(request):
-    return render_to_response("mips/logout.html",
-                              {"logout_url": users.create_logout_url("/")})
+    return render_to_response("mips/index.html", {'logged_in' : logged_in,
+                                                  'login_url' : users.create_login_url("/programs/"),
+                                                  'logout_url' : users.create_logout_url("/")
+                                                  })
 
 @Authenticated
 def programs(user, request):
     programs = UserProgram.all().filter("user =", user).fetch(10)
-    return render_to_response("mips/programs.html", {"programs": programs})
+    if not programs:
+        state_info = StateInfo()
+        state_info.put()
+        example = UserProgram(name="example", code=helpers.example_code(), user = user, state=state_info)
+        example.put()
+        programs.append(example)
+
+    exception = request.COOKIES.get('exception')
+    if not exception:
+        return render_to_response("mips/programs.html", {"programs": programs,
+                                                         "logout_url" : users.create_logout_url("/")})
+    else:
+        response  = render_to_response("mips/programs.html", {"programs": programs,
+                                                              "exception": exception,
+                                                              "logout_url" : users.create_logout_url("/")})
+        response.delete_cookie('exception')
+        return response
 
 @Authenticated
 def details(user, request, name):
-    current_program = UserProgram.all().filter("name =", name).filter("user =", user).get()
-    return render_to_response("mips/details.html", {'program' : current_program})
+    current_program = helpers.find_program(user, name)
+    return render_to_response("mips/details.html", {'program' : current_program,
+                                                    'logout_url' : users.create_logout_url("/")})
 
 @Authenticated
 def update(user, request):
     name = request.POST['name']
-    current_program = UserProgram.all().filter("name =", name).filter("user =", user).get()
+    current_program = helpers.find_program(user, name)
     current_program.code = request.POST['code']
     current_program.put()
-    return HttpResponseRedirect(reverse('mips.views.details', kwargs={'name':current_program.name}))
+    return HttpResponseRedirect(reverse('mips.views.details', kwargs={'name': current_program.name}))
 
 @Authenticated
 def add(user, request):
     name = request.POST['name']
-    state_info = StateInfo()
-    state_info.put()
-    new_program = UserProgram(name=name, code="", user=user, state=state_info)
-    new_program.put()
-    return HttpResponseRedirect(reverse('mips.views.programs'))
+    if not helpers.find_program(user, name):
+        state_info = StateInfo()
+        state_info.put()
+        new_program = UserProgram(name=name, code="", user=user, state=state_info)
+        new_program.put()
+        return HttpResponseRedirect(reverse('mips.views.programs'))
+    else:
+        msg = "Program with name '%s' already exists" % name
+        logging.info("EXCEPTION " + msg)
+        response = HttpResponseRedirect(reverse('mips.views.programs'))
+        response.set_cookie('exception', msg)
+        return response
 
 @Authenticated
 def delete(user, request):
     name = request.POST['name']
-    prog = UserProgram.all().filter("name =", name).filter("user =",user).get()
-    prog.delete();
-    return HttpResponseRedirect(reverse('mips.views.programs'))
+    prog = helpers.find_program(user, name)
+    if prog:
+        prog.delete();
+        return HttpResponseRedirect(reverse('mips.views.programs'))
+    else:
+        msg = "Cannot delete program with name '%s'<br /> because it does not exist." % name
+        response = HttpResponseRedirect(reverse('mips.views.programs'))
+        response.set_cookie('exception', msg)
+        return response
 
 @Authenticated
 def reset(user, request):
     name = request.POST['name']
-    query = UserProgram.all()
-    query.filter("name =", name)
-    prog = query.filter("user =", user).get()
+    prog = helpers.find_program(user, name)
+    prog.state.suspended = False
+    prog.state.put()
+    prog.put()
     return HttpResponseRedirect(reverse('mips.views.details',
                                         kwargs={'name':prog.name}))
 
@@ -108,4 +134,5 @@ def run(user, request, name):
                                       'output' : output})
         return HttpResponse(json_data, mimetype="application/javascript")
     else:
-        return HttpResponse("{'exception': '%s'}" % result['exception'], mimetype="application/javascript")
+        response = "{'exception': '%s', 'msg_list': %s}" % (result['exception'], result['msg_list'])
+        return HttpResponse(response, mimetype="application/javascript")
